@@ -1,9 +1,16 @@
 module github.com/vhive-serverless/vhive/utils/storage/go
 
-go 1.16
+go 1.19
 
 require (
 	github.com/aws/aws-sdk-go v1.40.16
 	github.com/go-redis/redis/v8 v8.11.2
 	github.com/sirupsen/logrus v1.8.1
+)
+
+require (
+	github.com/cespare/xxhash/v2 v2.1.1 // indirect
+	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
+	github.com/jmespath/go-jmespath v0.4.0 // indirect
+	golang.org/x/sys v0.0.0-20210423082822-04245dca01da // indirect
 )
