@@ -26,7 +26,7 @@ import (
 	"encoding/csv"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"math"
 	"os"
 	"path/filepath"
@@ -338,7 +338,7 @@ func loadMetrics(fileName string) (map[string]interface{}, error) {
 	}
 	defer jsonFile.Close()
 
-	byteValue, _ := ioutil.ReadAll(jsonFile)
+	byteValue, _ := io.ReadAll(jsonFile)
 	err = json.Unmarshal([]byte(byteValue), &result)
 	if err != nil {
 		return nil, err
